requirements.txt
fastapi
uvicorn
jinja2
python-jose[cryptography]
# passlib[bcrypt]
bcrypt>=4.0.0
passlib>=1.7.4
pymongo
python-multipart
strawberry-graphql
gunicorn
sqlmodel
pymysql
pymongo

